Modernize project from Java 1.5/Spring 3.1 to Java 17/Spring 6.2.3
Upgrade platform: Java 17, Spring 6.2.3, Spring Batch 5.2.2, JUnit 5,
HSQLDB 2.7.4, SLF4J/Logback replacing commons-logging/log4j.

Java modernization: Ledger as immutable record, java.time replacing
java.util.Date, BigDecimal for monetary fields, var, parameterized
logging, extracted helper methods, cleaned up signatures.

Spring Batch improvements: eliminated custom DAO layer (3 files) in
favor of JdbcBatchItemWriter, added LedgerProcessor for validation
and normalization, added skip/retry fault tolerance, migrated tests
to @SpringBatchTest, extracted shared base-batch-context.xml.

Observability: MDC context propagation, Micrometer metrics with
BatchMetricsLogger, enhanced job/step listeners with duration and
commit/rollback counts, structured logback pattern, rolling file
appender, full listener coverage on all three jobs.

Testing: expanded from 1 trivial test to 14 tests covering unit
tests for Ledger, LedgerMapper, LedgerRowMapper and integration
tests for all three batch jobs with proper assertions.

Added .gitignore and removed Eclipse IDE files from tracking.

+++ src/test/resources/schema-ddl.sql
@@ -5,8 +5,8 @@ create table ledger (
    chk_nbr VARCHAR(10) not null,
    chk_dt date,
    pymt_typ VARCHAR(50) not null,
-   dpst_amt double,
-   pymt_amt double,
+   dpst_amt DECIMAL(12,2),
+   pymt_amt DECIMAL(12,2),
    comments VARCHAR(100),
    constraint PID primary key (id)  
  );
